tegra194-flash-helper.sh: fix unique-pass check
The check for unique_pass=yes was accidentally instead checking for
uniquer_-guid=yes, which is not the name of the option. This led to always
passing --generic-pass to gen_luks_passphrase.py.

Also fix the derivation of ECID, which was erroneously passing ${chip}
as a string literal rather than the expanded value.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>