init-crypt-boot.sh: remove failing mount
The mounting of rootfs happens in platform-preboot-cboot.sh, after the
LUKS unlocking.

The logic to figure out the rootdev tries to mount the device with UUID
and fstype "crypto_LUKS", when it has already been mounted by dm-crypt in
/dev/mapper/crypt_root etc.

Add some echo statements and echo them to /dev/kmsg so they show up in
the logs.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>