platform-preboot-cboot: drop bad rootdev logic
The logic checking for "mayberoot" was not working properly and we ended
up with the rootdev still being /dev/mmcblk*. Add logic to match the
UUID, but this is not actually used. The mount is happening in the
"/etc/crypttab" code.

echo > /dev/kmsg so the messages actually show up in the logs.

Signed-off-by: Tim Orling <tim.orling@konsulko.com>